zsh: make prompt faster by only running vcs_info some of the time
To be honest, I rarely look at it, so it is unlikely that it being stale
will ever be a problem.

As you can see, some care is taken to run upon opening a new shell
(`LAST_COMMAND` will be empty), but not when the user just hits enter
(`LAST_COMMAND` would normally just keep its previous value, so we make
it "<unset>").