karabiner: make some tweaks for Spanish ISO keyboard
The </> key takes some of the space where left shift would usually be
and that happens to be where I usually try to hit shift so it was
driving me up the wall. So, map it to shift.

Likewise, the "ç" key is where part of the return key would usually be
and I was hitting that one all the time (especially because I
dual-purpose enter as control & enter). So, map it to enter.

The backtick key was "§" in Spanish ISO, which meant that I didn't know
how to type backtick or tilde... So, map it.

Finally, the upper part of the return key takes the place where the
\ and | key would normally be, and I haven't got a solution for that yet
(I briefly tried mapping the whole return key too it, but it was too
easy to accidentally get \ when I wanted return), so I will have to come
up with something else (probably holding down fn and hitting return, I
expect). But for now I've got a US external keyboard plugged in so
everything is good.