vim: fix double slashes in statusline
In fa55f82c9b47bd1af7ce9d006dcab57c4e4b2794 we replaced a hand-rolled
`substitute()` call with `fnamemodify()`, and attempted to fix a
regression by following up in f6a83f105801e0fda9a737ddc7793d8739f27365.

The problem is that `fnamemodify()` will return "~/" for the home
directory (note the trailing slash) but paths like "foo/bar" (no
trailing slash) elsewhere. So, f6a83f105 ended up fixing "foo/bar" (to
"foo/bar/") but breaking "~/" (changing it to "~//").

So in this commit, we strip the trailing slash, if present, then add it
back, so we always wind up with exactly one trailing slash no matter
what. This is more complicated that what we had prior to fa55f82c9b (ie.
just a simple `substitute()` call), but I think the `fnamemodify()` is
worth keeping around because of the path-shortening behavior we get from
the `:.` modifier.

Finally, I noticed we were using `fnamemodify()` without checking that
we're in a Vim compiled with support for it, so add a guard. It's
unlikely that we'll ever find ourselves running on such a Vim, but it's
just a matter of good hygiene.

Closes: https://github.com/wincent/wincent/issues/72