vim: fix TypeScript LSP usage in monorepo
In this monorepo:

    http://github.com/wincent/js

we have a top-level "tsconfig.json" file and various "package.json"
files at locations like:

    packages/something/package.json

This code:

    https://github.com/autozimu/LanguageClient-neovim/blob/b5dcd39f0845151e16c1f9096afd5d3265d1bc23/src/utils.rs#L47

is causing the wrong "rootPath" to be sent to the language server
implementations, which means that the "tsconfig.json" file is not
consulted, which causes problems with the type checking. For example, we
have "target" set to "esnext" in the config, but in Vim we see the
tsserver complaining about "Map" being an unknown type because it
doesn't see the config. This doesn't happen in VSCode, because it
correctly uses the "tsconfig.json" file at the project root.

Workaround here is to override "LanguageClient_rootMarkers", which
allows us to look first for "tsconfig.json", then ".flowconfig" (may as
well) and only finally "package.json".