vim: show LSP diagnostics signs in gutter
Taking special care here to not have things jump around when the column
hides/shows, and also trying to make the gutter line up correctly with
the red/green statusline indicator in as many cases as possible (eg.
when using vim-signature too).

Only thing I am not super found of here is that the 'signcolumn'
highlighting extends into the "EndOfBuffer" zone, even though the
"LineNr" do not.