vim: Boost deoplete file completion rank
Working in Alloy Editor where we have kebab-case file names I was
finding that filename completion in `require` and `import` statements
was painful with common words like "button" because it would end up
intermixing file results with other results, making it very hard to tab
to the next path component. Boost the "file" source rank seems to help a
lot.