vim: update to handle Neovim's "shada" replacement for "viminfo"
No idea why they did this, but whatever: 'viminfo' in Neovim is an alias
for 'shada'. Neovim reports that `has('viminfo')` is 0 (conversely Vim
says `has('shada')` is 0). The shada options are the same, but the
fileformat is different.

Because the formats are different we have to use different file names
for each.

I also removed the check for a viminfo file at ~/local/.vim/tmp; right
now I can't even remember what machine/OS I was running on that ever had
that as a default... let's see: ebd96c83c89247a5f4bd4917830af4272efaeec
says:

> vim: prefer ~/local/.vim/tmp over ~/.vim/tmp
>
> This is for my benefit when running out of an NFS-mounted home dir,
> where every time you create a file (including things like creating a
> commit message) you'll see this console noise:
>
>   "/tmp/hg-editor-rgAFDd.txt" 26L, 711C
>   Could not set security context for $HOME/.vim/tmp/undo/%tmp%hg-editor-rgAFDd.txt
>
> Basically, try the non-NFS ~/local symlink first, otherwise fall back.

So, it was corp-environment junk from FB.