zsh: define encrypted $GITHUB_ACCESS_TOKEN
As I am using it frequently enough of late in conjunction with
`github_changelog_generator -t $GITHUB_ACCESS_TOKEN`. As the token is
winding up in my shell history that way anyway, it's already on the
file-system, so storing it encrypted in the repo seems a not-terrible
thing to do.