doc: Update instructions for overriding `~/.gitconfig`
In ancient history I used to provide a hard-coded `~/.gitconfig` that
had stuff in it like my name ("user.name") and email ("user.email"),
but then people would clone this repo and wind up making commits as me
by mistake, so I added some runtime templating to make sure that
wouldn't happen.

In any case, the warning here is now a bit out of date. The right way to
override settings like "user.name" or "user.email" (or any other
setting) is to apply the changes to `~/.gitconfig.local`, so let's
document that.