vim: turn input-debouncing back on for Command-T
The portal repo has 3M files in it, so debouncing is probably necessary
again. Over the years, I started off without debouncing, then added it
when I started working in a repo that had 500K files. Then Command-T got
faster and I turned it off, even when working in a repo with > 1M files.
But in a 3M-file repo it is feeling laggy again. I'd like to make it
faster, but in the meantime this is the workaround to make things
livable.