vim: fix vcsjump "merge" functionality
This was working fine on my other machine but on this one it was
spamming a zsh warning* along the way. I had no idea that zsh was being
used as a shell anyway, so it seems like a good idea to make it boring
old "sh".

At the same time I made sure that we escape the arguments passed on to
the underlying tool, as I wouldn't want somebody to pass something like
"| rm -rf" or some other such thing.

[*]: zsh Can't add module parameter `modules': parameter already exists