vim: further improve diff view appearance
Doesn't look so great in Vim, where the prioritization of conceal
follows different rules (especially with respect to focused vs blurred
windows), but looks ok in Neovim. There is an underline being drawn
which looks kind of like 'cursorline' when moving inside added/changed
lines. Not sure how to turn that off.