zsh: propagate SIGINT to portool child processes more nicely
Before this commit, if you hit CTRL-C when running `portool tomcat`, for
instance, the script would die immediately and then pass the signal to
the child (in this case the tomcat instance). So you'd wind back at the
shell prompt but for a few more seconds tomcat would spew clean-up
messages to the console, conflicting with the prompt.

With this commit, we catch the signal, pass it to the child, wait for
the child to exit, then die off. In this way, the prompt doesn't clash
with the output of the child process.