zsh: add `portool {pause,resume,renice}`
"renice" is a bit standalone because the truth is it isn't effective.
Even at a "nice" level of 20, the build basically hoses the machine. So
I will probably delete it and that is why I haven't factored out the
duplication. Just wanted it to appear in the history though for the sake
of the record.