zsh: make `portool daily` really build portal-ee
I had been building "master", which won't do the right thing. Need to
use the "master-private" branch and run an Ant task to apply an overlay.

Using `:::` and `:::+` so that I can pass two arguments to each command
inside my `parallel` invocation, and running the Ant task as a separate
(serial) step because it only takes 30 seconds and this seems to strike
the right balance of simplicity and cleanness, even if it isn't ideal.