zsh: fix "portool {pause,resume}" edge case
This was working when I originally developed it because the process IDs
were 5 digits and there was no leading whitespace. Today, however, after
a reboot, my PIDs are 4-digit, which means that they had a leading space
which was breaking the `kill` command by producing arguments of the form
"- 1234" instead of "-1234".

Strip the leading whitespace.