fix(vim): rewind base16-vim submodule for compatibility with newer Vim
As per:

https://github.com/chriskempson/base16-vim/issues/197#issuecomment-492416355

this is the last known-good rev that will work in recent Vim containing:

https://github.com/vim/vim/commit/31b816042fca879b11965ddd75287732563ba698

PR 188 below is the one that needs to be backed out, and the colateral
damage is that we end up backing out PR 186 as well, which is actually a
useful one. Still, I deem this better than having to track a fork.

* roles/dotfiles/files/.vim/pack/bundle/start/base16-vim 2073e2d...5fc66da (3):
  < Merge pull request #186 from danielwe/master
  < Updates output
  < Merge pull request #188 from aaether32323/change_hi_function