refactor(vim)!: relocate submodules
BREAKING CHANGE: Although I moved them with `git mv` and everything
works fine, it annoys me when the name doesn't match the path in the
`.gitmodules` file, so I moved them under `.git/modules` and updated the
`gitdir:` links in the submodule `.git` files themselves. In a couple of
cases I had to edit the `config` file under `.git/modules/` too to make
it match.

Depending on when you cloned the repo, this may cause issues when you
`pull`. Cloning again will of course put everything into a pristine
up-to-date state if you don't want to fix any brokenness manually.