vim(feat): add `:Packadd` wrapper command
Instead of manually maintaining a list of plugins that have ftdetect
directories and therefore have to be loaded with `:packadd` instead of
`:packadd!`, make this helper that picks the right one automatically.

Of course, after having written all this, I realize it won't be
necessary, based on what `:h :packadd` says:

    If the filetype detection was not enabled yet (this
    is usually done with a "syntax enable" or "filetype
    on" command in your .vimrc file), this will also look
    for "{name}/ftdetect/*.vim" files.

ie. if I move my `:filetype` and `:syntax` commands to the bottom of the
file, and switch everything to use `:packadd!`, everything should just
work.

Going to commit this one for the historical record anyway.