fix(vim): status line left-hand-side colors not updating
Recently my LHS coloring stopped updating properly: it would no longer
update in response to text changes but it would in response to things
like moving between splits.

The cause is the `exists('#TextChangedI')` check was failing during
startup, probably due to my recent tweaks which affected load ordering.

From the `exists()` help:

      #event          autocommand defined for this event
      #event#pattern  autocommand defined for this event and
                      pattern (the pattern is taken
                      literally and compared to the
                      autocommand patterns character by
                      character)
      #group          autocommand group exists
      #group#event    autocommand defined for this group and
                      event.
      #group#event#pattern
                      autocommand defined for this group,
                      event and pattern.
      ##event         autocommand for this event is
                      supported.

I thought I was checking for "#event" being supported, but it seems I
was checking for an autocommand being defined for it. So, if a plugin
like Ultisnips (that defines such an autocommand) runs first, then our
`exists()` check returns true and everything works. But if it doesn't,
the check returns false and things are broken.

The fix is straightforward: `#event` needs to be `##event` instead.