feat(zsh): pass `fd` argument to `sk` instead of `bfs`
Previously we used the optional argument as a hard filter which would
search within a specific subdirectory. Now we instead pass it as an
argument to `sk`, to seed it with an initial query.

This will come in handy in the next commit, when I teach `jump` to fall
back to `fd` when a predefined directory hash isn't found.