fix(vim): broken override detection in compound file types
This is pretty lame:

    https://vi.stackexchange.com/questions/4893/how-to-make-autocmd-filetype-work-when-a-file-has-multiple-compound-dot-separa

Given a filetype like "typescript.tsx", an autocommand set up with any
of:

    autocmd FileType tsx ...
    autocmd FileType typescript ...
    autocmd FileType typescript,tsx ...

won't fire. Evidently it is doing an anchored search. This works:

    autocmd FileType tsx,typescript,typescript.tsx ...

Obviously not going to do that. Instead, run for all file types, and do
matching in the callback:

    autocmd FileType * ...

Seems preferable to the suggestion in the Stack Overflow post, which
could yield false positives.