feat(git): improve readability of "one"/"oneline" aliases
In a typical terminal split, I was having to deal with output like this:

    036b43af61e8 (HEAD -> wincent/LPS-99399/render-wrapper) LPS-99399 WIP (2 seconds ago) <Greg Hurrell>
    479600eade1c (tag: daily/portal/good/latest, tag: daily/portal/good/2019-08-20, upstream/master, upstream/HEAD, upstre

In terms of what is interesting here, the commit title is arguably the
most relevant, but it far too easily truncated over on the right (look
at the second commit; you can't see any of the title at all).

Swap the order from "decoration subject" to "subject decoration":

    036b43af61e8 LPS-99399 WIP (HEAD -> wincent/LPS-99399/render-wrapper) (5 minutes ago) <Greg Hurrell>
    479600eade1c LPS-99657 In the package, 'internal' should come first after the symbolic name of the bundle (tag: daily/

Not doing this for the "gr"/"graph"/"sl" aliases though, where the
"structural" information (ie. which branches are where) is arguably much
more relevant.

One little treachery to note: the "%d" format marker (for decoration)
always prints a leading space, which is why there is no whitespace after
the "%s" marker even though there is in the output.