feat(vim): always use hover preview in LanguageClient-neovim
Because otherwise it is annoying to not know whether "K" failed to show
anything because the output is only 1 line (in which case you have to
look down to the command line to see it) or because there was no output.

The glitch here is that it will show you an empty hover window if there
are 0 lines to display. I might file an upstream issue for that.