feat(zsh): tighten up portool helper checks
Instead of just checking for existence, check for:

- Having a simple "name" (ie. not a dotfile).
- Being executable.
- Being a file (ie. not a directory).

Because in the next commit I am going to stick some config files in a
directory.

Note the NodeJS API is a bit awkward: `accessSync` doesn't return a
boolean; it indicates success with `undefined` and an error by
`throw`-ing. `statSync` returns an object, or will throw for errors.