fix: don't try to run node during dotfiles install on Linux
The Karabiner stuff should only happen on Darwin (even there, it is a
bit fragile, as it assumes NodeJS is installed; will need to deal with
that at some point).

Note that even if the "register: karabiner_config" task gets skipped,
the variable is still set: it just won't have a "stdout" property, which
is why we check for the property in the template file (really, we
shouldn't even make the template file, but factoring that out would be
more work).

Closes: https://github.com/wincent/wincent/issues/79