feat(zsh)!: make `portool deploy` not clean by default
We already considered the `ALL` variable in the environment, but now we
consider `CLEAN` too, and it defaults to off. In my recent experience
I've found doing non-clean deploys works just fine.

One thing to bear in mind if this looks strange: the `-a` flag doesn't
mean "all" but the opposite; `-a` means "operate on this module only,
ignoring its dependencies". From the output of `gradlew --help`:

    -a, --no-rebuild          Do not rebuild project dependencies. [deprecated]

Deprecated here:

    https://github.com/gradle/gradle/issues/3077

But despite what it says on that ticket I think there's still some
benefit in using -a. Taking a typical module in liferay-portal (eg.
frontend-editor-alloyeditor-web), time with "-a" is about 6s vs time
without it of 1m57s.