feat(vim): implement "smart tab" functionality
The idea is to insert tabs on the left for indentation, and spaces on
the right for alignment. (In files that use hard tabs; in files with
'expandtab' set, we just use spaces.)

Based on the techniques in these plug-ins:

https://raw.githubusercontent.com/vim-scripts/Smart-Tabs/master/plugin/ctab.vim
https://github.com/vim-scripts/IndentTab/blob/master/autoload/IndentTab.vim

Those do quite a bit more than I really care to do (eg. they implement
":retab"-type functionality), so I think this will be enough for now.