feat(git): make tabs visible in diff views [experimental]
I have no idea how to write Perl any more so this is a total kludge but
I just want to see if it works.

Doing lots of whitespace-only (or mostly-whitespace) changes lately (see
https://github.com/wincent/liferay-portal/pull/64 for example) and it is
quite hard to see what is actually changing, even with a visual
difftool.

Swap out tabs in diffs for markers that match the listchars that I use
in Vim.

To see a diff without any of the "fanciness" you need to override the
pager:

    git --no-pager show # etc

One nice side-effect of this: it works even in `git add -p` (etc) so I'm
unlikely to ever add a tab by mistake.