feat(vim): add "compiler" file for ESLint
Instead of running :Lint, a slight improvement is:

    compiler eslint
    make

The error format here filters out the lines we don't care about and
additionally works with the default "stylish" formatter.

Once I've tested this a bit more I'll deal with the TODO comments and
then tweak :Lint to be a wrapper around this.

Got the idea for doing inline testing after the "finish" statement from
here:

https://www.reddit.com/r/vim/comments/d7vdbk/erroformat_playground/f1b578f/

Just run `:call wincent#debug#compiler()` to try it.