fix(vim): don't cd in wincent#liferay#deploy()
I am not sure if this code is the cause, but lately I've found Vim being
in the wrong directory (as though I had `'autochdir'` set, but I don't).
I suspect the `:cd` in here may be the cause, even though it supposedly
changes right back immediately with `:cd -`; maybe there's a race.

In any case, let's see if we can avoid that by `cd`-ing inside the
forked shell as opposed to prior to forking the shell.