perf(zsh): attempt to speed up `portool daily` by reducing logging
At least, visible logging.

The `exec` stuff is from:

https://stackoverflow.com/a/20564208

This allows us to redirect stderr and stdout to a file for the entire
duration of the subshell. Redirecting at the end (eg. `(...) 2>&1`)
doesn't work.

The `LOG_DESC` stuff is from:

https://stackoverflow.com/a/38462237

That allows us to print `~` in our messages instead of the full path.