feat(zsh): teach `git hub fetch` to work with N arguments
- 0 args: attempts to infer PR number.
- 1+ args: fetches all specified PRs.

Note the `--update-head-ok` hack from:

https://stackoverflow.com/a/19205680

Needed in order for the 0-arg case to work, otherwise Git will complain:

    fatal: Refusing to fetch into current branch refs/heads/pull/71 of
    non-bare repository

Obviously, this is super hacky. It won't touch the worktree, but it will
update the HEAD, so you are best advised to run `git status` to see what
the heck happened whenever you run a command like this.