fix(zsh): truncate log at start of each `portool daily` build
Based on inspection of the logs, I see that the previous setup was
just overwriting the old log in place without truncating it first (as
in, from byte 0 onwards); this could lead to confusing log output, where
the top of the file corresponded to the latest build, and the bottom to
previous builds (depending on their length, they might not get fully
overwritten).

It additionally meant that you couldn't meaningfully `tail` the log
during the build (because you'd just see the end of the file,
corresponding to the previous build).