chore(vim): update LanguageClient-neovim
I was seeing the opposite of what I was seeing in a63b0b39b71c891109ea:
auto-start is working on my personal laptop and not my work one (last
time it was working on my work one but not my personal one).

Last time, updating the plugin and re-running "install.sh" somehow fixed
it, but that hasn't worked this time. Oddly enough, I see these lines
when I open a buffer of the appropriate filetype and try to run an
LSP-powered action (like hitting `gd` to go to definition):

    [LC] Project root: /Users/greghurrell/code/portal/liferay-portal/modules/apps/asset/asset-taglib
    [LC] Language server is not running for: javascript.jsx

That first line is printed on open, suggesting the LanguageClient is
active; the second line is printed on hitting `gd`; it all works if I
do a `:LanguageClientStart` manually. Setting
`g:LanguageClient_autoStart` has no effect (as you would expect, because
it is on by default anyway).

* roles/dotfiles/files/.vim/pack/bundle/opt/LanguageClient-neovim 12e65e7...039bf7c (38):
  > Fix Makefile to work with CARGO_TARGET_DIR
  > Fix parsing of documentSymbol request
  > Skip text extracting when insert format is snippet
  > Format
  > Fix WorkspaceEditWithCursor definition
  > Try extract completion word from text_edit
  > Fix build warning
  > Fix pattern to match release version tag
  > Add an extra newline
  > Show relatedInformation in explainErrorAtPoint
  > Fix: install compare with the binary's version
  > Downgrade jsonrpc-core to fix #892
  > Add a clients_mutex for starting new language servers
  > (cargo-release) version 0.1.154
  > Keep consistent tag name
  > Simply update PATH on macOS CI as coreutils has already been installed
  > (cargo-release) version 0.1.153
  > Fixup lint
  > (cargo-release) version 0.1.152
  > Improve git tag to version conversion
  > (cargo-release) version 0.1.151
  > Remove unsupported target by cross
  > (cargo-release) version 0.1.150
  > Remove unsupported target by cross
  > (cargo-release) version 0.1.149
  > Revert adding SHA to git release
  > (cargo-release) version 0.1.148
  > Install coreutils on osx to produce SHA256
  > Don't highlight if range end is character 0
  > Suppress log for exit status 143 SIGTERM
  > Fix deoplete variable naming and conditional logic
  > Remove get_complete_position; default is bug-free
  > Fixup
  > Advertise publish_diagnostics related_information support
  > Fix error in FZFSinkLocation when filename contains ':'
  > Support filter diagnostics by severity
  > Added customizable virtual text prefix
  > Fix clippy::clone-on-copy