fix(vim): force LanguageClientStart in tsx/JSX files
I don't think I need the `LanguageClient_rootMarkers` to be spelled out
like that, but neither does it seem like that would be a problem. The
core bit is forcing a `LanguageClientStart` when we change the filetype,
because otherwise, we start one for "javascript"/"typescript"
automatically, but when the filetype updates to "javascript.jsx" or
"typescript.tsx" we don't, so an attempt to go-to-definition (eg. `gd`)
will fail.