feat(zsh): give "deployFast" another try
I was using this semi-successfully all day yesterday so going to give it
another shot. The main issue seems to be that when changes are detected,
we generate soy files during the build that the watcher then detects as
a second change, so it builds again. After two or three builds in a row
(not sure how many) it enters a steady state and stops. Not optimal, but
it appears that this only makes "deployFast" into "deployNotSoFast", not
"deployBroken".