fix(vim): correct path and naming problems in update process
In 268bca998c940 ("refactor(git)!: rename submodules", 2019-05-23) we
renamed submodules (formerly paths, now basenames), which means we need
to change the way we look up submodule config.

In 7718013f5080a ("refactor(vim): make it possible to easily disable
plugins", 2019-05-23) and 51a12a02a779e ('refactor(vim): move other
"opt" plugins under "bundle/opt"', 2019-05-23) we moved a bunch of
submodules around, which means we have to update a couple of paths.

In short, I broke a bunch of stuff on 2019-05-23 in the name of
progress, and because I have a very conservative update policy ("if it
ain't broke, don't fix it"), I am only discovering it now because I am
trying to do a bulk update (seeing as my tmux update broke Vim, may as
well get it all over with, I figure).