chore(zsh): reactivate binaries cache
I had this on because it maybe made things faster, then off because on
measuring I found it actually didn't make things faster (it would on a
slow network, but only if you had an up-to-date cache, and freshening it
on a slow network is not going to be quick). This morning, I was able to
do two successful builds with the cache off, but building the same rev
now (5.5 hours later) is failing. Turning cache on "fixes" it.