fix(zsh): make directory hash comparisons case-insensitive
Working on a different network today with work machine, noticed that
directory hashes aren't working, because my hostname is "lfrw044"
instead of the expected "LFRW044". I haven't rebooted since switching
networks, but my suspicion is that changing the network also ended up
changing the return value of `hostname -s` (in any case, `hostname` is
reporting `lfrw044.lan`; I can't remember what it reports on the other
network).

Forcing the comparisons to be case-insensitive with this Zsh trick:

https://stackoverflow.com/a/51391907/2103996

`$HOSTNAME:u` is uppercased, and `$HOSTNAME:l` is lowercased.