feat(zsh): remove scratch files more aggressively
I often wind up with Git-managed files in these directories that have
restrictive permissions and then print annoying prompts at exit:

    Removing scratch directory
    override r--r--r--  greghurrell/staff for /var/folders/yc/1x2y2qld1g95tvtjpd89hdrm0000gp/T/tmp.uaLPBTcR/.git/objects/59/eb23f5f60179abe5c5c3c5f7d431a66c437355? y
    override r--r--r--  greghurrell/staff for /var/folders/yc/1x2y2qld1g95tvtjpd89hdrm0000gp/T/tmp.uaLPBTcR/.git/objects/4b/825dc642cb6eb9a060e54bf8d69288fbee4904? y

Hopefully this won't bite me in the future. If there were a malicious
symlink in a scratch directory, the link would be deleted, not the
target. On the other hand, this document says that hard links to
directories are (still) not supported in Apple's APFS:

https://developer.apple.com/library/archive/documentation/FileManagement/Conceptual/APFS_Guide/FAQ/FAQ.html

> Directory hard links are not supported by Apple File System.

But that document is old, and I remember reading something somewhere
about how APFS implements snapshots using some kind of hard-link-like
trickery. Guess I'm just going to have to cross my fingers...