fix(vim): work around Emoji issues on Neovim
Troubleshooting process described here:

https://twitter.com/wincent/status/1238609346956857344

In short, I compared a large matrix of combinations:

- tmux built with --enable-utf8proc and without
- Inside and outside of tmux
- Inside and outside of Vim
- Bash vs Zsh
- iTerm vs Alacritty vs Kitty
- Various setting variants in iTerm
- Vim vs Neovim

Behavior is very complicated, so there were a number of confounding
factors (the biggest red-herring being that it seemed to happen only in
Zsh but not Bash -- but that was only because Zsh had an alias mapping
"vim" to "nvim").

In the end, this setting fixes display and editing in Neovim, and it
fixes display in Vim (editing worked there before and after).

Note that at the moment I am still running the tmux build that uses
utf8proc. That in itself was useful enough to get rid of the issues
where scrolling would corrupt the display. I didn't go back and check
tmux without utf8proc. Something to bear in mind for the future: if I
need to make a build of tmux with utf8proc, I can't use Homebrew,
because the `--enable-uft8proc` option was removed.