chore(zsh): {NO,}HIST_IGNORE_DUPS
Because the behavior on "up" is driving me batty. I'd *like* the history
file to be a literal record of the commands that I run, because that
means I can accurately figure out the most frequently run commands with:

    history 0 $HISTSIZE | awk '{a[$2]++}END{for(i in a){print a[i] " " i}}' | sort -rn | head

But I hit "up" hundreds of times a day. Better to have less-than-perfect
historical data that I look at once a year than be annoyed dozens of
times a day.