feat: implement variable propagation
Via horrible global state. But as implied in the comments, this is the
simplest way to have a "nice" DSL... ie:

    import {task} from '../../src/Fig';

    task('update terminfo files', ({command, path, variable}) => {
      for (const file of path.files('*.terminfo')) {
        command('tic', '-o', variable.string('terminfo_path'), file);
      }
    });

where the current aspect, variables etc are all implicit (inner calls
such as `command`, `path`, `variable` etc "know" about their context).

There might be other ways to do this, but they strike me as trickier to
implement.

One idea I'd like to leave the door open to is not passing the Fig
object into the task callback, which would mean the above would become:

    import {command, path, task, variable} from '../../src/Fig';

    task('update terminfo files', () => {
      for (const file of path.files('*.terminfo')) {
        command('tic', '-o', variable.string('terminfo_path'), file);
      }
    });

Which would be lighter weight (especially when there are a lot of
tasks), and would depend even more heavily on global state, because we
couldn't pass in special versions of `path`, `command` etc to the
callbacks.