refactor: simplify task callback calling convention
As mentioned in last commit. This makes us more dependent on global
state for communication, but it does make the DSL lighter, and it also
makes our type dependencies a bit less circular because we don't have to
pass the "Fig" type around all over the place. Overall, I think this is
an improvement.