feat(vim): add WIP of .editorconfig support
It works for the simple case shown here, which is changing the tabWidth
to 4 spaces (up from 2) in Markdown, JSON, JS, and TS files.

I prefer to do this minimally and locally without pulling in a large
editorconfig plugin like:

- https://github.com/editorconfig/editorconfig-vim
- https://github.com/sgur/vim-editorconfig

because I want to integrate it with my existing Liferay-specific stuff.

Plenty of unimplemented features though, and I don't intend to reach
feature parity with those plugins, so I might end up having to implement
an "escape hatch" here to blacklist folders when things don't play
nicely. Will cross that bridge when we come to it.