refactor(vim): get rid of 'shiftround' override
The reason why I need 'noshiftround' (actually the default), is that
otherwise my multi-line comments get mangled when dedenting/indenting in
files that use tabs for indentation (see dfa3c7950).

But, I don't actually see any reason why I ever set 'shiftround' in the
first place (in e3ea6b8f2ce8, back in 2010). As I said, 'noshiftround'
is the default, and it seems like a perfectly reasonable one. I am
leaving it in the settings explicitly so that it turns up in `git-blame`
output and people can find this explanation without having to trawl
through the log.