feat(vim, zsh): configure rg to search hidden files by default
Because I often want to find things in hidden directories (and not
just in repos like this one where I obviously need to edit a bunch of
dot-files), and I end up doing something like:

    :Ack -uu something

which has the unfortunate effect of not only searching through hidden
directories, but also ".gitignore"-d files, because of the way the `-u`
(`--unrestricted`) switch works:

- `-u` searches ignored files.
- `-uu` searches ignored and hidden files.
- `-uuu` searches ignored, hidden, and binary files.

So, if you're working in a folder with a `.gitignore`-d "node_modules"
folder with a zillion things in it, you'll find `rg` crawling through
that as well.

As such, you wind up doing something to compensate:

    :Ack -uu something -g !node_modules

Or switching to another tool entirely, because I know the thing I am
looking for is important enough to not be ignored but is tracked in the
Git repo:

    :VcsJump grep something

And both of those are of course terrible ideas because what you really
want is one simple way to search that does the right thing most of the
time without you having to constantly think about edge cases. So that's
what I am doing in this commit:

1. Make `--hidden` the default (will search hidden directories but still
   respects ".gitignore").
2. Exclude ".git", which is hidden and tends to be full of crap, but is
   not ignored so winds up polluting results. (Note: can add to this
   list if I encounter other problematic examples.)

If you want to override this temporarily:

- `--no-hidden` will undo the effect of `--hidden`.
- `-u` can still be used to search ignored files too.