fix(tmux): make double/triple-click behavior more consistent
The way we had it before, outside of copy mode, double or triple
clicking would enter copy mode and select a word or a line. But if you
double or triple clicked inside of copy mode, it would select a word or
a line, then copy it and leave copy mode.

This sounds like it might be a neat shortcut, but in practice, the
inconsistency between the behavior outside copy mode and the behavior
inside it is annoying. It's particularly frustrating if you scroll and
then absentmindedly double click: you end up copying, leaving copy mode,
and losing the scroll position.

In practice, the value of the copy-on-click shortcut is overrated: in
order to actually paste anywhere you need your hands on the keyboard, so
we may as well make it that the only way to copy is to hit return.