refactor(git): tweak a couple of aliases
To avoid mistakes like this one:

https://github.com/liferay/liferay-ckeditor/pull/56

In that repo, we have a submodule that is set up with an "ignore" value
of "all", which means that you don't see submodule changes in `git diff`
and you can't accidentally make changes with `git add`. At least, that's
the theory. You are supposed to use the provided "ck.sh" script in that
repo to do the updates, and that one uses `git add -f` under the cover
to do its work.

Turns out that `git add -u` also ends up adding submodule changes just
like `git add -f` would; "ouch", because it is super easy to do it by
accident and miss what you're doing because it isn't shown by `git
diff`.

We made a fix over there to catch mistakes in CI, but I can make a
change here that helps too: this commit makes `git staged` (and `git
both`, which is rarely used) show submodule commit changes, thus making
them harder to miss. Using "dirty" which will only show changes to the
commit that the submodule points at, which is what we care about in this
case (this is the only way to influence what is in a submodule checkout
from the contents of a commit).