fix(corpus): actually detect backticks
Given:

   foo `bar` baz

On finding first backtick, intent was to find second one, but we just
found the same one again due to our failure to add one to the index.
This broke code span detection, cause this to get improperly treated as
a shortcut link:

    foo `baz [bar] boom`

Same problem doesn't occur with links:

    foo [bar] baz

Because when we are at "[" and scan ahead looking for "]", there is no
possibility of finding "]" at "[" even without the index increment,
because they obviously don't match.