feat(fig): count skipped tasks
Normally we report "ok", "skipped" etc at the level of individual
operations, but here we were skipping over entire tasks and not counting
anything; I think it's reasonable to count here, otherwise the summary
output at the end may be misleading (eg. if you run with `--step` and
say "n" to every task, it would report a skipped total of 0, which
doesn't make any sense).