refactor(fig): set up path alias for "fig" -> "src"
This enables `tsc` to find modules from aspects like so:

    import '../../src/something.js';

instead of:

    import 'fig/something.js';

Normally I hate this kind of path hacking and aliasing because it
obfuscates where things really are.

But in this case aspects are supposed to be a DSL and somehow special.
"gd" in Vim still takes you to the right file (ie. via the LSP server).
Note that "gf" doesn't work after this change, and it didn't work before
it either, because when you are looking at "something.js", you actually
want to go to the source, which is "something.ts". So, "gd" works, but
"gf" doesn't.