fix(git): improve pager handling
Instead of applying a blanket pager to all commands, where it might not
work so great (eg. `git grep`: if you do `git grep non-existent-string`
it will fire up the pager and hold it open instead of exiting
immediately), apply it only to "diff-like" commands.

So that means we delete the `core.pager` config. For most commands, git
will use "less" by default, and because we set `LESS=-iFMRSX` in our
environment, it will use those options rather than overwriting `LESS`
with its preferred default (`-FSX`).

For the specific commands that produce (or can produce) diff-like
output, we set up our fancy pager just for those:

- `diff`
- `log` (ie. via `log -p`)
- `show`
- `whatchanged` (ie. via `whatchanged -p`, which is admittedly rare)

Then we have a number of aliases which wrap `diff` or `log`. Some of
those really aren't expected to produce diff-like output, so we opt out
of the fancy custom pager by doing `!git --paginate diff ...args`. This
means Git will still paginate (using `less` and `LESS`), but it won't
pay attention to the custom pagers defined in the `pager.<cmd>` config.

Example commands (aliases) that benefit from this:

- `git wd`
- `git graph`/`git gr`
- `git message`/`git msg`
- `git ten` (which I usually invoke with `oo`)
- `git oneline`/`git one` (which I usually invoke with `o`)
- `git sl`

Finally, the rarely used `git count` can opt out of pagination entirely
with `--no-pager`. I simplified the invocation to avoid the explicit
`sh` wrapper while I was at it.

Note that I changed my `LESS` export from `LESS=iFMRSX` to
`LESS=-iFMRSX` (both work), because the latter is what is used in the
`less` man page:

    For example, to avoid typing "less -options ..." each time less is
    invoked, you might tell csh:

       setenv LESS "-options"

One thing to note about all this: in testing I found that "-F" sometimes
didn't seem to work because I was printing less than a screenful but
`less` wasn't exiting. This was caused by long lines: any truncated line
will also cause `less` to stay open, independent of this change.

About the only remaining glitch is that `git diff` with a clean working
directory is still going to clear the screen before exiting. That may
just be something that I have to deal with.